backend/user-service/requirements.txt
Flask==3.0.3
PyJWT==2.8.0
PyMySQL==1.1.0